final hope

+++ app/src/test/java/mgarzon/createbest/productcatalog/ProductTest.java
@@ -20,7 +20,7 @@ public class ProductTest {
     public void checkProductPrice() {
         // I am making this test to fail
         Product aProduct = new Product("1", "DELL MONITOR", 180);
-        assertEquals("Check the price of the product",(double)180, aProduct.getPrice());//changed from string to doucle?
+        assertEquals("Check the price of the product",180.0, aProduct.getPrice());//changed from string to doucle?
     }
